compatibility with Tax Free Levels
