+++ src/main/java/amymialee/noenchantcap/MixinPlugin.java
@@ -1,0 +1,50 @@
+package amymialee.noenchantcap;
+
+import net.fabricmc.loader.api.FabricLoader;
+import org.objectweb.asm.tree.ClassNode;
+import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
+import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
+
+import java.util.List;
+import java.util.Set;
+
+public class MixinPlugin implements IMixinConfigPlugin {
+	@Override
+	public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
+		if (FabricLoader.getInstance().isModLoaded("taxfreelevels")) {
+			return !mixinClassName.endsWith("NECMixin_AnvilScreenHandler2");
+		}
+
+		return true;
+	}
+
+	@Override
+	public void onLoad(String mixinPackage) {
+
+	}
+
+	@Override
+	public String getRefMapperConfig() {
+		return null;
+	}
+
+	@Override
+	public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
+
+	}
+
+	@Override
+	public List<String> getMixins() {
+		return null;
+	}
+
+	@Override
+	public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+
+	}
+
+	@Override
+	public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+
+	}
+}
