+++ src/main/java/amymialee/noenchantcap/mixin/NECMixin_AnvilScreenHandler2.java
@@ -1,0 +1,39 @@
+package amymialee.noenchantcap.mixin;
+
+import amymialee.noenchantcap.NoEnchantCap;
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.screen.AnvilScreenHandler;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Constant;
+import org.spongepowered.asm.mixin.injection.ModifyConstant;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+@Mixin(AnvilScreenHandler.class)
+public class NECMixin_AnvilScreenHandler2 {
+    //Takes fair numbers of levels.
+    @Redirect(method = "onTakeOutput", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/player/PlayerEntity;addExperienceLevels(I)V"))
+    private void fairLevelCost(PlayerEntity instance, int levels) {
+        if (!NoEnchantCap.getConfig().fairLevelCost) {
+            instance.addExperienceLevels(levels);
+        } else {
+            instance.addExperience(-getLevelTotal(-levels));
+        }
+    }
+
+    private static int getLevelTotal(int level) {
+        int total = 0;
+        for (int i = 1; i <= level; i++) {
+            if (i >= 30) {
+                total += 112 + (i - 30) * 9;
+            } else if (i >= 15) {
+                total += 37 + (i - 15) * 5;
+            } else {
+                total += 7 + i * 2;
+            }
+        }
+        return total;
+    }
+}
